make Datei mit Dialog

+++ make-cb-tdd.sh
@@ -1,8 +1,21 @@
 #!/bin/bash
 
+dialog --backtitle "Make Cookbook TDD, 2015 BugBuster" --title "Frage" --yesno "Soll ein 'git pull' ausgeführt werden?" 5 50
+antwort=${?}
+
+if [ "$antwort" -eq "255" ]
+  then
+    echo "Abgebrochen"
+    exit 255
+fi
+
+if [ "$antwort" -eq "0" ]
+then
+  git pull
+fi
+
 rm -r cookbook-tdd/_book
 sync
-sleep 1
 
 gitbook build cookbook-tdd/
 if [ $? -eq 0 ]
@@ -14,7 +27,6 @@ else
 fi
 
 sync
-sleep 1
 html5validator --root cookbook-tdd/_book/ --ignore "gitbook/plugins"
 if [ $? -eq 0 ]
 then
@@ -25,7 +37,6 @@ else
 fi
 
 sync
-sleep 1
 gitbook pdf     cookbook-tdd/ ./cookbook-tdd.pdf
 if [ $? -eq 0 ]
 then
@@ -34,3 +45,4 @@ else
   echo "Problem beim generieren PDF"
   exit
 fi
+
